feat: simplify examples
