+++ src/components/CodeBlock/CodeBlock.tsx
@@ -8,7 +8,7 @@ import styles from './code-block.css?inline'
 import { IconCopySolid } from '../Icon'
 import { Button } from '~/components/Button/Button'
 
-type CodeBlockProps = { content: string; language: string; light?: boolean }
+type CodeBlockProps = { content: string; language: string; simple?: boolean }
 
 async function highlight(content: string, language: string) {
   // Escape angle brackets using HTML character entities
@@ -24,7 +24,7 @@ async function highlight(content: string, language: string) {
   return String(file)
 }
 
-export const CodeBlock = component$<CodeBlockProps>(({ content, language }) => {
+export const CodeBlock = component$<CodeBlockProps>(({ content, language, simple = false }) => {
   useStyles$(styles)
 
   const highlightedContent = useComputed$(async () => {
@@ -45,28 +45,38 @@ export const CodeBlock = component$<CodeBlockProps>(({ content, language }) => {
 
   return (
     <div class="border-gray-200 border-y border-x dark:border-gray-600 rounded-b-lg">
-      <div class="flex text-sm font-medium text-center justify-between text-gray-500 dark:text-gray-400 bg-gray-50 dark:bg-gray-700 border-gray-200 dark:border-gray-600 border-b">
-        <div class="inline-block p-2 px-3 text-gray-800 bg-gray-100 border-r border-gray-200 dark:text-white dark:bg-gray-800 dark:border-gray-600">
-          {language}
+      {!simple && (
+        <div class="flex text-sm font-medium text-center justify-between text-gray-500 dark:text-gray-400 bg-gray-50 dark:bg-gray-700 border-gray-200 dark:border-gray-600 border-b">
+          <div class="inline-block p-2 px-3 text-gray-800 bg-gray-100 border-r border-gray-200 dark:text-white dark:bg-gray-800 dark:border-gray-600">
+            {language}
+          </div>
+          <div class="flex justify-end">
+            <Button
+              onClick$={copy$}
+              color="light"
+              size="xs"
+              class="rounded-r-none border-t-0 border-r-0 border-b-0 rounded-l-none"
+              prefix={IconCopySolid}
+              title="Copy to clipboard"
+            >
+              {copyLabel}
+            </Button>
+          </div>
         </div>
-        <div class="flex justify-end">
-          <Button
-            onClick$={copy$}
-            color="light"
-            size="xs"
-            class="rounded-r-none border-t-0 border-r-0 border-b-0 rounded-l-none"
-            prefix={IconCopySolid}
-            title="Copy to clipboard"
-          >
-            {copyLabel}
-          </Button>
-        </div>
-      </div>
-      <pre dangerouslySetInnerHTML={highlightedContent.value} class={['p-4 overflow-auto', isCollapsed.value ? 'max-h-32' : 'max-h-none']} />
+      )}
+      <pre dangerouslySetInnerHTML={highlightedContent.value} class={['p-4 overflow-auto', isCollapsed.value ? 'max-h-48' : 'max-h-none']} />
 
-      <Button color="dark" size="md" full class="rounded-t-none border-t" onClick$={() => (isCollapsed.value = !isCollapsed.value)}>
-        {isCollapsed.value ? 'Expand code' : 'Collapse code'}
-      </Button>
+      {!simple && (
+        <Button
+          color="light"
+          size="md"
+          full
+          class="rounded-t-none border-t text-gray-800 bg-gray-100 border-gray-200 dark:text-white dark:bg-gray-800 dark:border-gray-600"
+          onClick$={() => (isCollapsed.value = !isCollapsed.value)}
+        >
+          {isCollapsed.value ? 'Expand code' : 'Collapse code'}
+        </Button>
+      )}
     </div>
   )
 })
