+++ src/components/__Preview/__Preview.tsx
@@ -68,6 +68,7 @@ export const Preview = component$<PreviewProps>(({ url, class: classNames, title
     <div>
       <h2 class="text-2xl font-bold text-gray-800 dark:text-gray-100 mb-3">{title}</h2>
       {props.description && <p class="text-gray-600 dark:text-gray-400 mb-4">{props.description}</p>}
+
       <div class="flex justify-between p-4 bg-gray-50 w-full border border-gray-200 rounded-t-xl dark:border-gray-600 dark:bg-gray-700">
         <ul>
           <li>
