+++ src/routes/examples/accordion/2.flush-accordion/index@examples.tsx
