+++ src/routes/examples/accordion/5.styling-accordion/index@examples.tsx
